+++ netket/nn/activation.py
@@ -12,13 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import functools
-from functools import wraps
+import jax as _jax
+from jax import numpy as _jnp
 
-import jax
-from jax import numpy as jnp
-
-from netket.utils import deprecated
+from netket.utils import deprecated as _deprecated
 
 from jax.nn import celu
 from jax.nn import elu
@@ -48,7 +45,7 @@ from jax.numpy import tanh
 from jax.numpy import cosh
 from jax.numpy import sinh
 
-from netket.jax import HashablePartial
+from netket.jax import HashablePartial as _HashablePartial
 
 
 def reim(f):
@@ -81,21 +78,21 @@ def log_cosh(x):
     """
     sgn_x = -2 * jnp.signbit(x.real) + 1
     x = x * sgn_x
-    return x + jnp.log1p(jnp.exp(-2.0 * x)) - jnp.log(2.0)
+    return x + _jnp.log1p(_jnp.exp(-2.0 * x)) - _jnp.log(2.0)
 
 
 def log_sinh(x):
     """
     Logarithm of the hyperbolic sine.
     """
-    return jax.numpy.log(jax.numpy.sinh(x))
+    return jnp.log(jnp.sinh(x))
 
 
 def log_tanh(x):
     """
     Logarithm of the hyperbolic tangent.
     """
-    return jax.numpy.log(jax.numpy.tanh(x))
+    return jnp.log(jnp.tanh(x))
 
 
 reim_selu = reim(selu)
@@ -106,18 +103,64 @@ r"""Returns the relu non-linearity, applied separately to the real and imaginary
 
 
 # TODO: DEPRECATION 3.1
-@deprecated("Deprecated. Use log_cosh instead")
+@_deprecated("Deprecated. Use log_cosh instead")
 def logcosh(x):
     return log_cosh(x)
 
 
 # TODO: DEPRECATION 3.1
-@deprecated("Deprecated. Use log_tanh instead")
+@_deprecated("Deprecated. Use log_tanh instead")
 def logtanh(x):
     return log_tanh(x)
 
 
 # TODO: DEPRECATION 3.1
-@deprecated("Deprecated. Use log_sinh instead")
+@_deprecated("Deprecated. Use log_sinh instead")
 def logsinh(x):
     return log_sinh(x)
+
+# TODO: Deprecation in 3.3 : Remove in the future
+_func_names = [
+    "celu",
+    "elu",
+    "gelu",
+    "glu",
+    "leaky_relu",
+
+    "log_sigmoid",
+    "log_softmax",
+    "normalize",
+    "relu",
+
+    "sigmoid",
+    "soft_sign",
+    "softmax",
+
+    "softplus",
+    "swish",
+    "silu",
+    "selu",
+    "hard_tanh",
+    "relu6",
+    "hard_sigmoid",
+    "hard_swish",
+]
+
+def _depmsg(func_name, new_module):
+    msg = f"""
+          `netket.nn.{func_name}` and `netket.nn.activation.{func_name}` are deprecated. 
+          Use `{new_module}.{func_name}` instead.
+          
+          Now that jax correctly supports complex numbers through it's activation functions,
+          please use them directly. `netket.nn.activation` will only contain activation
+          funtions specific to NetKet that are not available in `jax`.
+          
+          This warning will turn into an error in a future version of NetKet.
+          """
+    return msg
+
+for func_name in _func_names:
+    locals()[func_name] = _deprecated(_depmsg(func_name, new_module="[jax.nn/nn.activation]"), func_name)(getattr(_jax.nn, func_name))
+
+for func_name in ["cosh", "tanh", "sinh"]:
+    locals()[func_name] = _deprecated(_depmsg(func_name, new_module="jax.numpy"), func_name)(getattr(_jnp, func_name))